refactor: change composability of table for improved accessibility
